Improve not-found handling on working order creation. (#57)
Confirmations are not guaranteed to be processed immediately, and thus
may not be available immediately after order creation. It is
recommended to subscribe to TRADES:CONFIRMS or by polling /conrims
REST API. The confirms are only available for up to 1 minute, thus
implement polling inline/synchronously. Usually, the trades confirm
without a poll, or after just one poll.